コマンド毎にファイル分割、リファクタリング
