+++ Gemfile
@@ -11,4 +11,5 @@ gem "rufus-scheduler"
 group :development, :test do
   gem "rspec"
   gem "rubocop"
+  gem "pry"
 end
